Add the generation definition (but results not drawn yet)

+++ Environment.h
@@ -51,7 +51,7 @@ class Environment
     //                            Accessors: getters
     // =======================================================================
 	
-    int* getFecondity(void) const;
+    double* getFecondity(void) const;
 
 
     // =======================================================================
